chore(meta): move FetchAddU64Response utils to separate file
